+++ packages/isaacscript-common/src/features/customStage/customStageGridEntities.ts
@@ -16,9 +16,11 @@ import {
 import { calculateStageType } from "../../functions/stage";
 import { vectorEquals } from "../../functions/vector";
 import { CustomStage } from "../../interfaces/private/CustomStage";
-import { TrapdoorDestination } from "../../types/TrapdoorDestination";
 import { isCustomGridEntity } from "../customGridEntity";
-import { spawnCustomTrapdoor } from "../customTrapdoor/exports";
+import {
+  spawnCustomTrapdoor,
+  spawnCustomTrapdoorToVanilla,
+} from "../customTrapdoor/exports";
 import { DEFAULT_BASE_STAGE } from "./exports";
 import v from "./v";
 
@@ -203,20 +205,22 @@ export function convertVanillaTrapdoors(
 
   removeGridEntity(gridEntity, true);
 
-  // - If we are on the first floor of a custom stage, then the destination will be the second floor
-  //   of the custom stage. (e.g. Caves 1 to Caves 2)
-  // - If we are on the second floor of a custom stage, then the destination will be the vanilla
-  //   floor equivalent to 2 floors after the floor used as a basis for the custom stage.
-  const baseStage =
-    customStage.baseStage === undefined
-      ? DEFAULT_BASE_STAGE
-      : customStage.baseStage;
-  const vanillaNextStage = (baseStage + 2) as LevelStage;
-  const vanillaNextStageType = calculateStageType(vanillaNextStage);
-  const destination: TrapdoorDestination = v.run.firstFloor
-    ? [customStage.name, 2]
-    : [vanillaNextStage, vanillaNextStageType];
-  spawnCustomTrapdoor(gridEntity.Position, destination);
+  if (v.run.firstFloor) {
+    // If we are on the first floor of a custom stage, then the destination will be the second floor
+    // of the custom stage. (e.g. Caves 1 to Caves 2)
+    spawnCustomTrapdoor(gridEntity.Position, customStage.name, 2);
+  } else {
+    // If we are on the second floor of a custom stage, then the destination will be the vanilla
+    // floor equivalent to 2 floors after the floor used as a basis for the custom stage.
+    const baseStage =
+      customStage.baseStage === undefined
+        ? DEFAULT_BASE_STAGE
+        : customStage.baseStage;
+    const stage = (baseStage + 2) as LevelStage;
+    const stageType = calculateStageType(stage);
+
+    spawnCustomTrapdoorToVanilla(gridEntity.Position, stage, stageType);
+  }
 }
 
 /**